requirements.txt
#
# This file is autogenerated by pip-compile with python 3.8
# To update, run:
#
#    pip-compile
#
cycler==0.11.0
    # via matplotlib
joblib==1.0.1
    # via
    #   -r requirements.in
    #   scikit-learn
kiwisolver==1.4.2
    # via matplotlib
lightgbm==3.2.1
    # via -r requirements.in
matplotlib==3.4.3
    # via
    #   -r requirements.in
    #   seaborn
numpy==1.19.5
    # via
    #   -r requirements.in
    #   lightgbm
    #   matplotlib
    #   pandas
    #   scikit-learn
    #   scipy
    #   seaborn
pandas==1.3.2
    # via
    #   -r requirements.in
    #   seaborn
pillow==9.1.0
    # via matplotlib
pyparsing==3.0.8
    # via matplotlib
python-dateutil==2.8.2
    # via
    #   matplotlib
    #   pandas
pytz==2022.1
    # via pandas
pyyaml==5.4.1
    # via -r requirements.in
scikit-learn==0.24.2
    # via
    #   -r requirements.in
    #   lightgbm
scipy==1.8.0
    # via
    #   lightgbm
    #   scikit-learn
    #   seaborn
seaborn==0.11.2
    # via -r requirements.in
six==1.16.0
    # via python-dateutil
threadpoolctl==3.1.0
    # via scikit-learn
wheel==0.37.1
    # via lightgbm
